fixed bugs
commented out the drawHands method as in its current state it breaks everything.

fixed the issue where only the first cards image would load (the method used to make the card matrix didn't remove the card after returning it(drawCard), so it just filled up with the same card(changed drawCard to takeCard to fix)), also made the decks shuffle through the shuffle method in the collections interface.
(if whenever you load up the game it seems that some cards don't load, its because of an issue with the file system, it doesn't break anything but we still need to fix it soon).

fixed and simplified patrons in both the logic and splendorPanel classes by converting the patrons list in the splendorPanel class from an array to an array list to allow for easier shuffling, as array Lists have built in shuffling methods through the collections interface, (also removed the CurrPatreons variable from the logic class, as it was unnecessary).