fixed bugs and added the end game check
fixed bug relating to token calculations (sometimes the methods that would take away tokens specifically would not account for discounts, leading to it outputting negative amounts of tokens), also fixed an issue relating to the card matrix display (since it only displayed the first 4 cards, whenever a card was taken, it would shift the order off all the other cards which looked weird, fixed by having a separate matrix 2d array variable in logic that just replaces cards with the next in line when they are removed).

added a end game detection to the program which displays text if you are on the last round, and switches to the end Panel when you end the last round, also simplified and combined some methods in logic relating to this as having it all in one method was much simpler.

finally, reorganized the logic methods to be more consistent with each other (all get methods are in one place, for example).